+++ src/main/services/pingEnvironmentsJob.ts
