+++ src/main/services/validateOAuthPermission.ts
